+++ src/WebApi.Authentication/Consts.cs
@@ -1,0 +1,6 @@
+namespace WebApi.Authentication;
+
+internal class Consts
+{
+	public const string SegregationClaim = "segregation-key";
+}