+++ src/WebApi.Authentication/DependencyInjection/IWebApiAuthenticationServiceCollection.cs
@@ -33,3 +33,23 @@ public interface IWebApiAuthenticationServiceCollection<TApiSecret>
 	IWebApiAuthenticationServiceCollection<TApiSecret> AddApiSecretRepository<TRepository>(Func<IServiceProvider, TRepository> repositoryFactory)
 		where TRepository : class, IApiSecretRepository<TApiSecret>;
 }
+
+/// <summary>
+///     Custom ServiceCollection for adding repositories to the WebApi.Authentication configuration.
+/// </summary>
+/// <typeparam name="TApiSecret">Type of ApiSecret. Must inherit SegregatedApiSecret</typeparam>
+/// <typeparam name="TKey">Type of key to use for repository segregation</typeparam>
+public interface IWebApiAuthenticationServiceCollection<TApiSecret, out TKey>
+	where TApiSecret : SegregatedApiSecret<TKey>
+	where TKey : IParsable<TKey>
+{
+	/// <summary>
+	///     Adds a repository to the service collection for storing ApiSecrets using a custom ApiSecret type and a segregation
+	///     key for the underlying repositories.
+	/// </summary>
+	/// <param name="repositoryFactory">Factory method to create the repository</param>
+	/// <typeparam name="TRepository">Type of repository</typeparam>
+	/// <returns>service collection</returns>
+	IWebApiAuthenticationServiceCollection<TApiSecret, TKey> AddSegregatedApiSecretRepository<TRepository>(Func<IServiceProvider, TKey, TRepository> repositoryFactory)
+		where TRepository : class, IApiSecretRepository<TApiSecret>;
+}