+++ src/WebApi.Authentication/DependencyInjection/WebApiAuthenticationServiceCollection.cs
@@ -1,13 +1,10 @@
 namespace WebApi.Authentication.DependencyInjection;
 
-internal class WebApiAuthenticationServiceCollection<TApiSecret> : IWebApiAuthenticationServiceCollection<TApiSecret>
+internal class WebApiAuthenticationServiceCollection<TApiSecret> : BaseWebApiAuthenticationServiceCollection<TApiSecret>, IWebApiAuthenticationServiceCollection<TApiSecret>
 	where TApiSecret : ApiSecret
 {
-	private readonly IServiceCollection _services;
-
-	public WebApiAuthenticationServiceCollection(IServiceCollection services)
+	public WebApiAuthenticationServiceCollection(IServiceCollection services) : base(services)
 	{
-		_services = services;
 	}
 
 	public IWebApiAuthenticationServiceCollection<TApiSecret> AddApiSecretRepository<TRepository>()
@@ -49,8 +46,41 @@ internal class WebApiAuthenticationServiceCollection<TApiSecret> : IWebApiAuthen
 		AddAdapterRepository();
 		return this;
 	}
+}
 
-	private void AddAdapterRepository()
+internal class WebApiAuthenticationServiceCollection<TApiSecret, TKey> : BaseWebApiAuthenticationServiceCollection<TApiSecret>, IWebApiAuthenticationServiceCollection<TApiSecret, TKey>
+	where TApiSecret : SegregatedApiSecret<TKey>
+	where TKey : IParsable<TKey>
+{
+	public WebApiAuthenticationServiceCollection(IServiceCollection services) : base(services)
+	{
+	}
+
+	public IWebApiAuthenticationServiceCollection<TApiSecret, TKey> AddSegregatedApiSecretRepository<TRepository>(Func<IServiceProvider, TKey, TRepository> repositoryFactory)
+		where TRepository : class, IApiSecretRepository<TApiSecret>
+	{
+		if (_services.Any(sd => sd.ServiceType == typeof(IApiSecretRepository)))
+		{
+			throw new InvalidOperationException("An ApiSecret repository has already been registered.");
+		}
+
+		_services.AddSingleton<IApiSecretRepository<TApiSecret>>(provider => new SegregatedApiSecretRepositoryAdapter<TKey, TApiSecret, TRepository>(provider, repositoryFactory));
+		AddAdapterRepository();
+		return this;
+	}
+}
+
+internal abstract class BaseWebApiAuthenticationServiceCollection<TApiSecret>
+	where TApiSecret : class, IApiSecret
+{
+	protected readonly IServiceCollection _services;
+
+	protected BaseWebApiAuthenticationServiceCollection(IServiceCollection services)
+	{
+		_services = services;
+	}
+
+	protected void AddAdapterRepository()
 	{
 		_services.AddSingleton<IApiSecretRepository>(provider => new ApiSecretRepositoryAdapter<TApiSecret>(provider.GetRequiredService<IApiSecretRepository<TApiSecret>>()));
 	}