+++ tests/WebApi.Authentication.UnitTests/CustomApiSecret.cs
@@ -1,6 +1,5 @@
 namespace WebApi.Authentication.UnitTests;
 
-public class CustomApiSecret : ApiSecret
-{
-	public string Role { get; init; } = "";
-}
+public class CustomApiSecret : ApiSecret;
+
+public class CustomSegregatedApiSecret : SegregatedApiSecret<Guid>;