+++ tests/WebApi.Authentication.UnitTests/DependencyInjectionTests/WebApiAuthenticationServiceCollectionTests.cs
@@ -172,10 +172,71 @@ public class WebApiAuthenticationServiceCollectionTests
 
 		Assert.Equal("An ApiSecret repository has already been registered.", ex.Message);
 	}
+
+	[Fact]
+	public void AddSegregatedApiSecretRepositoryFactory_AddedTwice_Throws()
+	{
+		// Arrange
+		var services = new ServiceCollection();
+		var webApiAuthenticationServiceCollection = services.AddSegregatedApiSecretAuthentication<SegregatedApiSecret<Guid>, Guid>(TestHelper.CreateConfiguration());
+		webApiAuthenticationServiceCollection.AddSegregatedApiSecretRepository<FakeRepository<SegregatedApiSecret<Guid>>>((_, _) =>
+			new FakeRepository<SegregatedApiSecret<Guid>>());
+
+		// Act && Assert
+		var ex = Assert.Throws<InvalidOperationException>(() =>
+			webApiAuthenticationServiceCollection.AddSegregatedApiSecretRepository<FakeRepository<SegregatedApiSecret<Guid>>>((_, _) =>
+				new FakeRepository<SegregatedApiSecret<Guid>>()));
+
+		Assert.Equal("An ApiSecret repository has already been registered.", ex.Message);
+	}
+
+	[Fact]
+	public void AddSegregatedApiSecretRepositoryGeneric_CustomType_AddedWithCustomType()
+	{
+		// Arrange
+		var services = new ServiceCollection();
+		var webApiAuthenticationServiceCollection = services.AddSegregatedApiSecretAuthentication<CustomSegregatedApiSecret, Guid>(TestHelper.CreateConfiguration());
+
+		// Act
+		webApiAuthenticationServiceCollection.AddSegregatedApiSecretRepository<FakeRepository<CustomSegregatedApiSecret>>((_, _) => new FakeRepository<CustomSegregatedApiSecret>());
+
+		var provider = services.BuildServiceProvider();
+
+		// Assert
+		var genericRepository = provider.GetRequiredService<IApiSecretRepository<CustomSegregatedApiSecret>>();
+		var plainRepository = provider.GetRequiredService<IApiSecretRepository>();
+		Assert.NotNull(genericRepository);
+		Assert.NotNull(plainRepository);
+		Assert.IsType<SegregatedApiSecretRepositoryAdapter<Guid, CustomSegregatedApiSecret, FakeRepository<CustomSegregatedApiSecret>>>(genericRepository);
+		Assert.IsType<ApiSecretRepositoryAdapter<CustomSegregatedApiSecret>>(plainRepository);
+	}
+
+	[Fact]
+	public void AddSegregatedApiSecretRepositoryGeneric_DefaultType_AddedWithDefaultType()
+	{
+		// Arrange
+		var services = new ServiceCollection();
+		var webApiAuthenticationServiceCollection = services.AddSegregatedApiSecretAuthentication<SegregatedApiSecret<Guid>, Guid>(TestHelper.CreateConfiguration());
+
+		// Act
+		webApiAuthenticationServiceCollection.AddSegregatedApiSecretRepository<FakeRepository<SegregatedApiSecret<Guid>>>((_, _) =>
+			new FakeRepository<SegregatedApiSecret<Guid>>());
+
+
+		var provider = services.BuildServiceProvider();
+
+		// Assert
+		var genericRepository = provider.GetRequiredService<IApiSecretRepository<SegregatedApiSecret<Guid>>>();
+		var plainRepository = provider.GetRequiredService<IApiSecretRepository>();
+		Assert.NotNull(genericRepository);
+		Assert.NotNull(plainRepository);
+		Assert.IsType<SegregatedApiSecretRepositoryAdapter<Guid, SegregatedApiSecret<Guid>, FakeRepository<SegregatedApiSecret<Guid>>>>(genericRepository);
+		Assert.IsType<ApiSecretRepositoryAdapter<SegregatedApiSecret<Guid>>>(plainRepository);
+	}
 }
 
 file class FakeRepository<T> : IApiSecretRepository<T>
-	where T : ApiSecret
+	where T : class, IApiSecret
 {
 	public ValueTask<T?> GetByClaimsAsync(ICollection<Claim> claims, CancellationToken cancellationToken = default)
 	{